+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Mercury
 
 # Mercury.Common v1.0.5.3
-# Mercury.Reservations v1.0.1.1
+# Mercury.Reservations v1.0.1.2
 
 ## 2021-08-28
 1. [ADD] Adds validation that requires ExpiredAt for rooms, must be at least one more day than today
+2. [ADD] Adds RoomFixture for creating mock data
+3. [ADD] Adds filter active to only get the active rooms
 
 ## 2021-08-21
 1. Adds validation uniqueness title and Id for Rooms